refactor(agentic): extract posting and verification nodes into standalone modules (#93, #94)
